# openapi.yaml
openapi: 3.0.0
info:
  title: Gestão de Horários Escolares
  version: 1.0.0
  description: API REST para gerir alunos, professores, salas, aulas e horários.
servers:
  - url: http://localhost:3000
paths:
  /alunos:
    get:
      summary: Lista todos os alunos
      responses:
        '200':
          description: Lista de alunos
    post:
      summary: Cria um novo aluno
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nome:
                  type: string
                email:
                  type: string
                curso:
                  type: string
      responses:
        '201':
          description: Aluno criado
  /alunos/{id}:
    get:
      summary: Obtém aluno por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Aluno encontrado
        '404':
          description: Não encontrado
    put:
      summary: Atualiza aluno por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nome:
                  type: string
                email:
                  type: string
                curso:
                  type: string
      responses:
        '200':
          description: Aluno atualizado
    delete:
      summary: Remove aluno por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Aluno removido

  /professores:
    get:
      summary: Lista todos os professores
      responses:
        '200':
          description: Lista de professores
    post:
      summary: Cria um novo professor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nome:
                  type: string
                email:
                  type: string
                departamento:
                  type: string
      responses:
        '201':
          description: Professor criado
  /professores/{id}:
    get:
      summary: Obtém professor por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Professor encontrado
        '404':
          description: Não encontrado
    put:
      summary: Atualiza professor por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nome:
                  type: string
                email:
                  type: string
                departamento:
                  type: string
      responses:
        '200':
          description: Professor atualizado
    delete:
      summary: Remove professor por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Professor removido

  /salas:
    get:
      summary: Lista todas as salas
      responses:
        '200':
          description: Lista de salas
    post:
      summary: Cria uma nova sala
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nome:
                  type: string
                capacidade:
                  type: integer
      responses:
        '201':
          description: Sala criada
  /salas/{id}:
    get:
      summary: Obtém sala por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Sala encontrada
        '404':
          description: Não encontrada
    put:
      summary: Atualiza sala por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nome:
                  type: string
                capacidade:
                  type: integer
      responses:
        '200':
          description: Sala atualizada
    delete:
      summary: Remove sala por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Sala removida

  /aulas:
    get:
      summary: Lista todas as aulas
      responses:
        '200':
          description: Lista de aulas
    post:
      summary: Cria uma nova aula
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                disciplina:
                  type: string
                id_professor:
                  type: integer
      responses:
        '201':
          description: Aula criada
  /aulas/{id}:
    get:
      summary: Obtém aula por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Aula encontrada
        '404':
          description: Não encontrada
    put:
      summary: Atualiza aula por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                disciplina:
                  type: string
                id_professor:
                  type: integer
      responses:
        '200':
          description: Aula atualizada
    delete:
      summary: Remove aula por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Aula removida

  /horarios:
    get:
      summary: Lista todos os horários
      responses:
        '200':
          description: Lista de horários
    post:
      summary: Cria um novo horário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                dia_semana:
                  type: string
                hora_inicio:
                  type: string
                hora_fim:
                  type: string
                id_sala:
                  type: integer
                id_professor:
                  type: integer
                id_aula:
                  type: integer
      responses:
        '201':
          description: Horário criado
  /horarios/{id}:
    get:
      summary: Obtém horário por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Horário encontrado
        '404':
          description: Não encontrado
    put:
      summary: Atualiza horário por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                dia_semana:
                  type: string
                hora_inicio:
                  type: string
                hora_fim:
                  type: string
                id_sala:
                  type: integer
                id_professor:
                  type: integer
                id_aula:
                  type: integer
      responses:
        '200':
          description: Horário atualizado
    delete:
      summary: Remove horário por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Horário removido